+++ src/services/leafletMapDefaults.js
@@ -1,4 +1,4 @@
-angular.module("leaflet-directive").factory('leafletMapDefaults', function ($q, leafletHelpers) {
+angular.module('ui-leaflet').factory('leafletMapDefaults', function ($q, leafletHelpers) {
     function _getDefaults() {
         return {
             keyboard: true,