+++ test/unit/boundsDirectiveSpec.js
@@ -7,7 +7,7 @@
 describe('Directive: bounds', function() {
     var $compile, $rootScope, leafletData, bounds, scope;
 
-    beforeEach(module('leaflet-directive'));
+    beforeEach(module('ui-leaflet'));
     beforeEach(inject(function(_$compile_, _$rootScope_, _leafletData_){
         $compile = _$compile_;
         $rootScope = _$rootScope_;